requirements.txt
anyio==3.7.1
Flask==3.0.0
requests==2.30.0
telebot==0.0.5
beautifulsoup4==4.12.2
python-telegram-bot==20.3
python-dotenv==1.0.0